Refactor buffer flags and resource binding for improved clarity and functionality
- Renamed `BufferFlags::CPU_COHERENT` to `BufferFlags::CPU_READABLE` to better reflect its purpose in managing CPU access to buffers.
- Updated documentation and method signatures across the codebase to align with the new buffer flag naming.
- Removed deprecated `read_buffer_coherent` methods from backends and replaced them with `read_to_cpu` for consistent CPU read operations.
- Enhanced the `DeviceCapabilities` struct to include support for zero-copy storage readback, improving performance on compatible backends.
- Adjusted integration tests to reflect changes in buffer flag usage and ensure correctness in CPU read operations.

+++ src/task_graph/analysis.rs
@@ -1259,7 +1259,7 @@ mod tests {
                 node(
                     "write",
                     i,
-                    vec![(range(0, i as u64 * 128, 128), NodeAccess::Write)],
+                    vec![(range(0, i * 128, 128), NodeAccess::Write)],
                     1,
                 )
             })
